requirements.txt
six>=1.10.0
numpy>=1.12.1
tqdm>=4.11.2
torchvision>=0.1.8
